code_assessment_platform/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
dj-rest-auth>=5.0
django-allauth>=0.61
mysqlclient>=2.2
PyMySQL>=1.1
python-dotenv>=1.0
requests>=2.31
